Remove show.go - handle errors better for invalid roles
